Alternative raster generation from LAS
This is already covered, but this will take a LAS file and submit the job in a parallelized fashion to ROGER (or another HPC environment with the correct prerequisites), and produce GeoTIFF rasters for each LAS file.  New commands to manipulate the data should be added to the "template.json" file as necessary.